Redid input
I have redone the way input is used for both mouse and keyboard input.
Now objects are notified when there is a mouse input or a keyboard input
and then can chose what to do with it. This input includes information
about the event. With this new system, I have also added in mouse
presses, mouse releases, and mouse scrolling (although scrolling is a
bit wonky because of integer rounding for now)